fibonacci

+++ Fibonacci/src/App.java
@@ -6,7 +6,7 @@ public class App
 	{
 		Fibo f = new Fibo();
 		System.out.println(f.getFibonacci());
-		for (long i = 0; i < 91; i++)
+		for (int i = 0; i < 91; i++)
 		{
 			System.out.println(f.next().getFibonacci());
 		}
